backend/requirements.txt
# Core dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# ML/AI
torch==2.1.0
transformers==4.35.0
accelerate==0.24.1
bitsandbytes==0.41.1
optimum==1.14.0
sentencepiece==0.1.99
protobuf==4.25.0

# Data processing
numpy==1.24.3
pandas==2.0.3

# Async & Caching
redis==5.0.1
aioredis==2.0.1
asyncio==3.4.3

# Auth & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
bcrypt==4.1.1

# Monitoring & Logging
prometheus-client==0.19.0
python-json-logger==2.0.7
sentry-sdk==1.38.0

# HTTP & WebSocket
websockets==12.0
httpx==0.25.1
aiohttp==3.9.0

# Utilities
python-dotenv==1.0.0
pyyaml==6.0.1
click==8.1.7
tqdm==4.66.1

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.1
